+++ bf_opencore/celery.py
@@ -25,6 +25,7 @@ from django.conf import settings
 from bf_opencore.models import Asset
 import json
 import math
+from bf_opencore.models.viper import ViperWebhookRequest, ViperWebhookResponse, ViperAsset
 
 # register signals
 from bf_opencore import signals
@@ -32,12 +33,7 @@ from bf_opencore import signals
 logger = logging.getLogger(__name__)
 
 # Create a Celery instance (sometimes called an app)
-# Based on Celery documentation found at
-#
-# Celery docs
 # http://docs.celeryproject.org/en/latest/getting-started/next-steps.html
-# http://docs.celeryproject.org/en/latest/reference/celery.html#celery.Celery.config_from_object
-# http://docs.celeryproject.org/en/latest/userguide/configuration.html
 celery_app = celery.Celery('bf_opencore')
 
 # Use Django settings module for configuration
@@ -53,60 +49,6 @@ celery_app.config_from_object(
     namespace='CELERY',
     silent=False,
 )
-
-@dataclass
-class ViperWebhookRequest:
-    """Data for a viper webhook."""
-    callback: str
-    since: str # iso8601 
-    before: str # iso8601
-    page: int
-    page_size: int
-
-class ViperAsset:
-    """Data for a viper asset."""
-    id: int
-    network_segment: str
-    cpe: str
-    role: str
-    upstream_api: str # asset endpoint url: {BASE_URL}/api/assets/{id}/
-    hostname: str
-    mac_address: str
-    serial_number: str
-    location: dict[str, str]
-    status: str
-    vendorID: int
-
-    def __init__(self, asset: Asset):
-        self.id = asset.id
-        self.network_segment = asset.groups.first().name
-        self.cpe = asset.custom_fields.get('cpe').value
-        self.role = asset.category
-        self.upstream_api = f"{settings.BASE_URL}/api/assets/{asset.id}/"
-        self.hostname = asset.hostname
-        self.mac_address = asset.mac_address
-        self.serial_number = asset.serial_number
-
-    def __init__(self, asset: Asset):
-        self.id = asset.id
-        self.name = asset.name
-        self.ip_address = asset.ip_address
-        self.mac_address = asset.mac_address
-        self.vendor = asset.manufacturer
-        self.model = asset.model
-        self.serial_number = asset.serial_number
-        self.udi = asset.udi
-
-@dataclass
-class ViperWebhookResponse:
-    """Response for a viper webhook."""
-    items: list[ViperAsset]
-    page: int
-    page_size: int
-    total: int
-    total_pages: int
-    next_page: str | None # url to the next page: {BASE_URL}/api/assets/?page={page+1}&page_size={page_size}
-    previous_page: str | None # url to the previous page: {BASE_URL}/api/assets/?page={page-1}&page_size={page_size}
 
 
 @celery_app.task
